Main/VideoRecordingAndEncoding : build fix for windows platform.
svn:sync-xref-src-repo-rev: r80098

+++ src/VBox/Main/include/MachineImpl.h
@@ -1,4 +1,4 @@
-/* $Id: MachineImpl.h 42839 2012-08-16 09:38:41Z noreply@oracle.com $ */
+/* $Id: MachineImpl.h 42841 2012-08-16 10:30:24Z noreply@oracle.com $ */
 /** @file
  * Implementation of IMachine in VBoxSVC - Header.
  */
@@ -421,10 +421,10 @@ public:
     STDMETHOD(COMSETTER(VideoCaptureEnabled))(BOOL  u8VideoRecEnabled);
     STDMETHOD(COMGETTER(VideoCaptureFile))(BSTR * ppChVideoRecFilename);
     STDMETHOD(COMSETTER(VideoCaptureFile))(IN_BSTR pChVideoRecFilename);
-    STDMETHOD(COMGETTER(VideoCaptureWidth))(uint32_t *u32VideoRecHorzRes);
-    STDMETHOD(COMSETTER(VideoCaptureWidth))(uint32_t u32VideoRecHorzRes);
-    STDMETHOD(COMGETTER(VideoCaptureHeight))(uint32_t *u32VideoRecVertRes);
-    STDMETHOD(COMSETTER(VideoCaptureHeight))(uint32_t u32VideoRecVertRes);
+    STDMETHOD(COMGETTER(VideoCaptureWidth))(ULONG *u32VideoRecHorzRes);
+    STDMETHOD(COMSETTER(VideoCaptureWidth))(ULONG u32VideoRecHorzRes);
+    STDMETHOD(COMGETTER(VideoCaptureHeight))(ULONG *u32VideoRecVertRes);
+    STDMETHOD(COMSETTER(VideoCaptureHeight))(ULONG u32VideoRecVertRes);
     STDMETHOD(COMGETTER(MonitorCount))(ULONG *monitorCount);
     STDMETHOD(COMSETTER(MonitorCount))(ULONG monitorCount);
     STDMETHOD(COMGETTER(Accelerate3DEnabled))(BOOL *enabled);
